Fix SceneObject rotation inconsistency.

+++ SW/environment/src/hu/nik/project/environment/objects/SceneObject.java
@@ -33,7 +33,7 @@ public abstract class SceneObject<T> implements Serializable {
     abstract public T getObjectType();
 
     public ScenePoint getBasePosition() { return basePosition; }
-    public double getRotation() {
+    public int getRotation() {
         return rotation;
     }
